+++ tests/testthat/test-renv.R
@@ -25,24 +25,24 @@ test_that("Ensure docker cache dir", {
   expect_true(fs::dir_exists(target))
 })
 
-test_that("Digest primary manifest", {
-  result <- renv_digest_primary_manifest()
+test_that("Digest primary lockfile", {
+  result <- renv_digest_primary_lockfile()
 
   expect_type(result, "character")
   expect_identical(nchar(result), 32L)
 })
 
-test_that("Digest cached manifest", {
-  result <- renv_digest_cached_manifest()
+test_that("Digest cached lockfile", {
+  result <- renv_digest_cached_lockfile()
 
   expect_type(result, "character")
   expect_identical(nchar(result), 32L)
 })
 
-test_that("Manage cached manifest", {
-  result <- renv_manage_cached_manifest()
+test_that("Manage cached lockfile", {
+  result <- renv_manage_cached_lockfile()
 
-  target <- .env_renv_manifest_cached()
+  target <- .env_renv_lockfile_cached()
 
   expect_identical(result, target)
   expect_true(fs::file_exists(target %>%