+++ NEWS.md
@@ -1,3 +1,12 @@
+# renv2docker 0.0.0.9002
+
+Renamed functions & devtools
+
+- Renamed
+- Added package dependency: `devtools`
+
+--------------------------------------------------------------------------------
+
 # renv2docker 0.0.0.9001
 
 Fixed export